requirements.txt
emoji
indic-nlp-library
pandas
scikit-learn
transformers
torch
-e .